+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Added support for query scopes declared with Laravel's `#[Scope]` attribute, including protected methods
+* Added support for the `#[CollectedBy]` attribute so queries return the model's custom collection
 
 ## Version 1.9.0 (2026-05-20)
 * Added `DriverRegistry` so custom drivers can be registered with `register()`